observability/backend/requirements.txt
fastapi==0.123.0
uvicorn==0.38.0
sqlalchemy==2.0.44
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.2.1
python-jose[cryptography]==3.3.0
slowapi==0.1.9
boto3==1.35.0
httpx==0.27.0
google-auth==2.29.0
requests==2.32.5
pydantic[email]==2.12.5
